fix: keep the FastWAM Gate in FP32 master weights
The Gate adopted the frozen model dtype, so its bfloat16 parameters
discarded every Adam update below half a BF16 ULP. Replaying one real Adam
step from the global_step_200 optimizer moments moves only 5.55% of Gate
elements; its routing head moves 1.17% of its weights and 0% of its bias.
The Gate never left initialization across 2000 optimizer steps even though
gradients and Adam moments were present throughout.

The Gate is now built in FP32. `DirectKVAttention` already casts the stored
K/V banks to the query dtype, so the read-only tap is unchanged.

Gate K/V materialization no longer keys on the Gate parameter dtype. An
FP32 Gate would otherwise have upcast every replayed layer, doubling replay
memory and breaking the declared per-sample and tier byte limits; it now
uses the configured storage dtype and `DirectKVAttention` upcasts one layer
at a time.

`partition_fastwam_trainable_parameters` fails closed on any
reduced-precision trainable so this cannot silently return.

See `docs/BF16_PARAMETER_UPDATE_LOSS.md`.

+++ rlinf/models/embodiment/wam_policy/__init__.py
@@ -386,6 +386,7 @@ def _validate_critic_build_config(cfg) -> bool:
 def get_model(cfg, torch_dtype):
     """Build the composite policy from explicit FastWAM/OpenPi sub-configs."""
 
+    import torch
     from fastwam.adapters import (
         RegimeLoRAConfig,
         inject_action_dit_lora,
@@ -533,7 +534,12 @@ def get_model(cfg, torch_dtype):
         actor.action_expert,
         lora_config,
     )
-    gate = GateTransformer(gate_config).to(dtype=torch_dtype)
+    # The Gate is trainable, so it stays in FP32 rather than adopting the frozen
+    # model dtype. A BF16 Gate discards every Adam update below half a BF16 ULP,
+    # which is where this Gate's updates land; see
+    # `docs/BF16_PARAMETER_UPDATE_LOSS.md`. `DirectKVAttention` already casts the
+    # stored K/V banks to the query dtype, so the read-only tap still works.
+    gate = GateTransformer(gate_config).to(dtype=torch.float32)
 
     critic = None
     critic_feature_config = None
